+++ projects/wallet-template/src/background/createBackgroundRpc.ts
@@ -15,7 +15,7 @@ import { UserSignedExtensionName } from "../types/UserSignedExtension"
 import { createClient } from "@polkadot-api/substrate-client"
 import { getObservableClient } from "@polkadot-api/observable-client"
 import { filter, firstValueFrom, map, mergeMap, take } from "rxjs"
-import { createTx } from "./tx/create-tx"
+import { createTx } from "./tx-helper/create-tx"
 import * as pjs from "./pjs"
 import { Bytes, Variant } from "@polkadot-api/substrate-bindings"
 import { InPageRpcSpec } from "../inpage/types"
