+++ projects/wallet-template/src/background/tx-helper/descriptors.ts
