+++ projects/wallet-template/src/background/tx-helper/index.ts
