+++ projects/wallet-template/src/background/tx-helper/signed-extensions/chain/CheckMetadataHash.ts
