+++ projects/wallet-template/src/background/tx-helper/signed-extensions/chain/CheckNonce.ts
