+++ projects/wallet-template/src/background/tx-helper/signed-extensions/index.ts
