+++ projects/wallet-template/src/background/tx-helper/signed-extensions/internal-types.ts
