+++ projects/wallet-template/src/background/tx-helper/signed-extensions/user/index.ts
