+++ projects/wallet-template/src/background/tx-helper/signed-extensions/utils.ts
