+++ projects/wallet-template/src/background/tx-helper/types.ts
